+++ rollup.config.js
@@ -1,39 +1,17 @@
 import path from 'path'
-import commonjs from '@rollup/plugin-commonjs'
-import json from '@rollup/plugin-json'
-import nodeResolve from '@rollup/plugin-node-resolve'
-import typescript from '@rollup/plugin-typescript'
 import builtinModules from 'builtin-modules'
-import { terser } from 'rollup-plugin-terser'
 import pkg from './package.json'
+import { compact, sourcemap, plugins } from './rollup.setup'
 
-const compact = !process.env.ROLLUP_WATCH
-const sourcemap = !compact
-
-const plugins = [
-  json({ preferConst: true }),
-  nodeResolve({ mainFields: ['module', 'jsnext:main', 'main'] }),
-  commonjs(),
-  typescript(),
-  compact && terser(),
-]
-
-export default [
-  {
-    external: [...Object.keys(pkg.dependencies), ...builtinModules, 'vscode'],
-    input: `src/${path.basename(pkg.main, '.js')}.ts`,
-    output: {
-      compact,
-      exports: 'named',
-      file: pkg.main,
-      format: 'cjs',
-      sourcemap,
-    },
-    plugins,
+export default {
+  external: [...Object.keys(pkg.dependencies), ...builtinModules, 'vscode'],
+  input: `src/${path.basename(pkg.main, '.js')}.ts`,
+  output: {
+    compact,
+    exports: 'named',
+    file: pkg.main,
+    format: 'cjs',
+    sourcemap,
   },
-  {
-    input: 'preview.js',
-    output: { compact, file: 'lib/preview.js', format: 'iife', sourcemap },
-    plugins,
-  },
-]
+  plugins: plugins(),
+}