+++ src/__mocks__/fs.ts
@@ -1,7 +1,0 @@
-module.exports = {
-  ...jest.requireActual('fs'),
-
-  // Mock file R/W
-  unlink: jest.fn((_, cb) => cb(null)),
-  writeFile: jest.fn((_, __, cb) => cb(null)),
-}