+++ src/marp-cli.test.ts
@@ -1,11 +1,10 @@
-import fs from 'fs'
 import { tmpdir } from 'os'
 import path from 'path'
 import * as marpCliModule from '@marp-team/marp-cli'
 import { workspace } from 'vscode'
 import * as marpCli from './marp-cli'
+import { textEncoder } from './utils'
 
-jest.mock('fs')
 jest.mock('vscode')
 
 const setConfiguration: (conf?: Record<string, unknown>) => void = (
@@ -101,7 +100,7 @@ describe('#createWorkFile', () => {
     expect(workFile.path).toEqual('/tmp/clean.md')
 
     await workFile.cleanup()
-    expect(fs.unlink).not.toHaveBeenCalled()
+    expect(workspace.fs.delete).not.toHaveBeenCalled()
   })
 
   it('creates tmpfile to same directory of file when passed a dirty file', async () => {
@@ -115,21 +114,21 @@ describe('#createWorkFile', () => {
       workFile.path.startsWith(path.join('/tmp', '.marp-vscode-tmp'))
     ).toBe(true)
 
-    expect(fs.writeFile).toHaveBeenCalledWith(
-      workFile.path,
-      'example',
-      expect.any(Function)
+    expect(workspace.fs.writeFile).toHaveBeenCalledWith(
+      expect.objectContaining({ fsPath: workFile.path }),
+      textEncoder.encode('example')
     )
 
     await workFile.cleanup()
-    expect(fs.unlink).toHaveBeenCalledWith(workFile.path, expect.any(Function))
+    expect(workspace.fs.delete).toHaveBeenCalledWith(
+      expect.objectContaining({ fsPath: workFile.path }),
+      expect.any(Object)
+    )
   })
 
   it('creates tmpfile to workspace root when failed creating to same dir', async () => {
     // Simulate that creation to same directory is not permitted
-    ;(fs as any).writeFile.mockImplementationOnce((_, __, cb) =>
-      cb(new Error())
-    )
+    jest.spyOn(workspace.fs, 'writeFile').mockRejectedValueOnce(new Error())
 
     jest
       .spyOn(workspace, 'getWorkspaceFolder')