+++ src/utils.ts
@@ -1,6 +1,8 @@
 import { AbortController } from 'abort-controller'
-import nodeFetch from 'node-fetch'
-import { TextDocument, workspace } from 'vscode'
+import fetchPonyfill from 'fetch-ponyfill'
+import { TextDocument, Uri, workspace } from 'vscode'
+
+export const _fetchPonyfillInstance = fetchPonyfill()
 
 interface FetchOption {
   timeout?: number
@@ -30,7 +32,8 @@ export const fetch = (url: string, { timeout = 5000 }: FetchOption = {}) => {
   const controller = new AbortController()
   const timeoutCallback = setTimeout(() => controller.abort(), timeout)
 
-  return nodeFetch(url, { signal: controller.signal })
+  return _fetchPonyfillInstance
+    .fetch(url, { signal: controller.signal })
     .then((res) => {
       if (!res.ok) throw new Error(`Failured fetching ${url} (${res.status})`)
       return res.text()
@@ -49,3 +52,15 @@ export const mathTypesettingConfiguration = () => {
   )
   return conf ?? 'katex'
 }
+
+export const textEncoder = new globalThis.TextEncoder()
+export const textDecoder = new globalThis.TextDecoder()
+
+export const readFile = async (target: Uri) =>
+  textDecoder.decode(await workspace.fs.readFile(target))
+
+export const writeFile = (target: Uri, text: string) =>
+  workspace.fs.writeFile(target, textEncoder.encode(text))
+
+export const unlink = (target: Uri) =>
+  workspace.fs.delete(target, { useTrash: false })